My first leetcode problem solved duplicate.
I have solved a duplicate by using the Set() method to store an arrays value and then used a for loop to iterate through the given array which has a conditional statement that checks if the variable, seen (which has the Set() method), already has the number that its currently on. If it does, it returns the first duplicate. If not, then it adds the numbers to set and continues. If no duplicate is in the array, it returns -1